+++ AssemblyInterpret/Program.cs
@@ -40,9 +40,9 @@ class Program
         memory.PrintMemory();
 
 
-        int strAddress = memory.GetGlobalVar("str")!.Address;
+        uint strAddress = memory.GetGlobalVar("str")!.Address;
         Console.WriteLine();
-        for (int i = 0; i < 5; i++)
+        for (byte i = 0; i < 5; i++)
         {
             Console.Write("{0}", (char)memory.ReadByte(strAddress + i));
         }
